Add templates to HTTP server, listener and session, make dummy metricshandler which doesn't do anything
