+++ src/communication/http/session.hpp
@@ -12,13 +12,19 @@
 #pragma once
 
 #include <deque>
+#include <functional>
 #include <memory>
 #include <optional>
+#include <string>
 #include <variant>
 
+#include <spdlog/spdlog.h>
+#include <boost/asio/bind_executor.hpp>
 #include <boost/asio/dispatch.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/asio/strand.hpp>
+#include <boost/beast/core/buffers_to_string.hpp>
+#include <boost/beast/core/stream_traits.hpp>
 #include <boost/beast/core/tcp_stream.hpp>
 #include <boost/beast/http.hpp>
 #include <boost/beast/ssl.hpp>
@@ -26,12 +32,22 @@
 #include <json/json.hpp>
 
 #include "communication/context.hpp"
+#include "communication/http/handler.hpp"
+#include "utils/logging.hpp"
 #include "utils/variant_helpers.hpp"
 
 namespace memgraph::communication::http {
 
-class Session : public std::enable_shared_from_this<Session> {
+namespace {
+void LogError(boost::beast::error_code ec, const std::string_view what) {
+  spdlog::warn("HTTP session failed on {}: {}", what, ec.message());
+}
+}  // namespace
+
+template <class TRequestHandler>
+class Session : public std::enable_shared_from_this<Session<TRequestHandler>> {
   using tcp = boost::asio::ip::tcp;
+  using std::enable_shared_from_this<Session<TRequestHandler>>::shared_from_this;
 
  public:
   template <typename... Args>
@@ -39,23 +55,119 @@ class Session : public std::enable_shared_from_this<Session> {
     return std::shared_ptr<Session>{new Session{std::forward<Args>(args)...}};
   }
 
-  void Run();
+  void Run() {
+    if (auto *ssl = std::get_if<SSLSocket>(&stream_); ssl != nullptr) {
+      try {
+        boost::beast::get_lowest_layer(*ssl).expires_after(std::chrono::seconds(30));
+        ssl->handshake(boost::asio::ssl::stream_base::server);
+      } catch (const boost::system::system_error &e) {
+        spdlog::warn("Failed on SSL handshake: {}", e.what());
+        return;
+      }
+    }
+
+    // run on the strand
+    boost::asio::dispatch(strand_, [shared_this = shared_from_this()] { shared_this->DoRead(); });
+  }
 
  private:
   using PlainSocket = boost::beast::tcp_stream;
   using SSLSocket = boost::beast::ssl_stream<boost::beast::tcp_stream>;
 
-  explicit Session(tcp::socket &&socket, ServerContext &context);
+  explicit Session(tcp::socket &&socket, ServerContext &context)
+      : stream_(CreateSocket(std::move(socket), context)), strand_{boost::asio::make_strand(GetExecutor())} {}
 
-  std::variant<PlainSocket, SSLSocket> CreateSocket(tcp::socket &&socket, ServerContext &context);
+  std::variant<PlainSocket, SSLSocket> CreateSocket(tcp::socket &&socket, ServerContext &context) {
+    if (context.use_ssl()) {
+      ssl_context_.emplace(context.context_clone());
+      return Session::SSLSocket{std::move(socket), *ssl_context_};
+    }
 
-  void OnWrite(boost::beast::error_code ec, size_t bytes_transferred);
+    return Session::PlainSocket{std::move(socket)};
+  }
 
-  void DoRead();
-  void OnRead(boost::beast::error_code ec, size_t bytes_transferred);
+  void OnWrite(boost::beast::error_code ec, size_t bytes_transferred) {
+    boost::ignore_unused(bytes_transferred);
 
-  void DoClose();
-  void OnClose(boost::beast::error_code ec);
+    if (ec) {
+      close_ = true;
+      return LogError(ec, "write");
+    }
+
+    if (close_) {
+      DoClose();
+      return;
+    }
+
+    res_ = nullptr;
+
+    DoRead();
+  }
+
+  void DoRead() {
+    req_ = {};
+
+    ExecuteForStream([this](auto &&stream) {
+      boost::beast::get_lowest_layer(stream).expires_after(std::chrono::seconds(30));
+
+      boost::beast::http::async_read(
+          stream, buffer_, req_,
+          boost::asio::bind_executor(strand_, std::bind_front(&Session::OnRead, shared_from_this())));
+    });
+  }
+  void OnRead(boost::beast::error_code ec, size_t bytes_transferred) {
+    boost::ignore_unused(bytes_transferred);
+
+    if (ec == boost::beast::http::error::end_of_stream) {
+      DoClose();
+      return;
+    }
+
+    if (ec) {
+      return LogError(ec, "read");
+    }
+
+    auto async_write = [this](boost::beast::http::response<boost::beast::http::string_body> msg) {
+      ExecuteForStream([this, &msg](auto &&stream) {
+        // The lifetime of the message has to extend
+        // for the duration of the async operation so
+        // we use a shared_ptr to manage it.
+        auto sp = std::make_shared<boost::beast::http::response<boost::beast::http::string_body>>(std::move(msg));
+
+        // Store a type-erased version of the shared
+        // pointer in the class to keep it alive.
+        res_ = sp;
+        // Write the response
+        boost::beast::http::async_write(
+            stream, *sp, boost::asio::bind_executor(strand_, std::bind_front(&Session::OnWrite, shared_from_this())));
+      });
+    };
+
+    // handle request
+    HandleRequest(std::move(req_), async_write);
+  }
+
+  void DoClose() {
+    std::visit(utils::Overloaded{[this](SSLSocket &stream) {
+                                   boost::beast::get_lowest_layer(stream).expires_after(std::chrono::seconds(30));
+
+                                   // Perform the SSL shutdown
+                                   stream.async_shutdown(
+                                       boost::beast::bind_front_handler(&Session::OnClose, shared_from_this()));
+                                 },
+                                 [](PlainSocket &stream) {
+                                   boost::beast::error_code ec;
+                                   stream.socket().shutdown(tcp::socket::shutdown_send, ec);
+                                 }},
+               stream_);
+  }
+  void OnClose(boost::beast::error_code ec) {
+    if (ec) {
+      LogError(ec, "close");
+    }
+
+    // At this point the connection is closed gracefully
+  }
 
   auto GetExecutor() {
     return std::visit(utils::Overloaded{[](auto &&stream) { return stream.get_executor(); }}, stream_);
@@ -71,6 +183,7 @@ class Session : public std::enable_shared_from_this<Session> {
   boost::beast::flat_buffer buffer_;
   boost::beast::http::request<boost::beast::http::string_body> req_;
   std::shared_ptr<void> res_;
+  TRequestHandler handler;
   boost::asio::strand<boost::beast::tcp_stream::executor_type> strand_;
   bool close_{false};
 };
